+++ src/main/java/net/geforcemods/securitycraft/blockentities/KeyPanelBlockEntity.java
@@ -1,6 +1,7 @@
 package net.geforcemods.securitycraft.blockentities;
 
 import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.CustomizableBlockEntity;
 import net.geforcemods.securitycraft.api.ILockable;
 import net.geforcemods.securitycraft.api.IPasswordProtected;
@@ -8,22 +9,19 @@ import net.geforcemods.securitycraft.api.Option;
 import net.geforcemods.securitycraft.api.Option.BooleanOption;
 import net.geforcemods.securitycraft.api.Option.IntOption;
 import net.geforcemods.securitycraft.blocks.KeyPanelBlock;
-import net.geforcemods.securitycraft.inventory.GenericBEMenu;
 import net.geforcemods.securitycraft.misc.ModuleType;
+import net.geforcemods.securitycraft.network.client.OpenCheckPasswordScreen;
+import net.geforcemods.securitycraft.network.client.OpenSetPasswordScreen;
 import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.MenuProvider;
-import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraftforge.network.NetworkHooks;
+import net.minecraftforge.network.PacketDistributor;
 
 public class KeyPanelBlockEntity extends CustomizableBlockEntity implements IPasswordProtected, ILockable {
 	private String passcode;
@@ -80,39 +78,14 @@ public class KeyPanelBlockEntity extends CustomizableBlockEntity implements IPas
 
 	@Override
 	public void openPasswordGUI(Player player) {
-		if (getPassword() != null) {
-			if (player instanceof ServerPlayer serverPlayer) {
-				NetworkHooks.openGui(serverPlayer, new MenuProvider() {
-					@Override
-					public AbstractContainerMenu createMenu(int windowId, Inventory inv, Player player) {
-						return new GenericBEMenu(SCContent.CHECK_PASSWORD_MENU.get(), windowId, level, worldPosition);
-					}
-
-					@Override
-					public Component getDisplayName() {
-						return KeyPanelBlockEntity.super.getDisplayName();
-					}
-				}, worldPosition);
-			}
-		}
-		else {
-			if (getOwner().isOwner(player)) {
-				if (player instanceof ServerPlayer serverPlayer) {
-					NetworkHooks.openGui(serverPlayer, new MenuProvider() {
-						@Override
-						public AbstractContainerMenu createMenu(int windowId, Inventory inv, Player player) {
-							return new GenericBEMenu(SCContent.SET_PASSWORD_MENU.get(), windowId, level, worldPosition);
-						}
-
-						@Override
-						public Component getDisplayName() {
-							return KeyPanelBlockEntity.super.getDisplayName();
-						}
-					}, worldPosition);
-				}
-			}
+		if (!level.isClientSide) {
+			if (getPassword() != null)
+				SecurityCraft.channel.send(PacketDistributor.PLAYER.with(() -> (ServerPlayer) player), new OpenCheckPasswordScreen(worldPosition));
 			else {
-				PlayerUtils.sendMessageToPlayer(player, new TextComponent("SecurityCraft"), Utils.localize("messages.securitycraft:passwordProtected.notSetUp"), ChatFormatting.DARK_RED);
+				if (getOwner().isOwner(player))
+					SecurityCraft.channel.send(PacketDistributor.PLAYER.with(() -> (ServerPlayer) player), new OpenSetPasswordScreen(worldPosition));
+				else
+					PlayerUtils.sendMessageToPlayer(player, new TextComponent("SecurityCraft"), Utils.localize("messages.securitycraft:passwordProtected.notSetUp"), ChatFormatting.DARK_RED);
 			}
 		}
 	}