+++ src/main/java/net/geforcemods/securitycraft/blockentities/UsernameLoggerBlockEntity.java
@@ -5,7 +5,6 @@ import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.ILockable;
 import net.geforcemods.securitycraft.api.Option;
 import net.geforcemods.securitycraft.api.Option.IntOption;
-import net.geforcemods.securitycraft.inventory.GenericBEMenu;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.network.client.ClearLoggerClient;
 import net.geforcemods.securitycraft.network.client.UpdateLogger;
@@ -14,17 +13,13 @@ import net.geforcemods.securitycraft.util.ITickingBlockEntity;
 import net.geforcemods.securitycraft.util.ModuleUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.chat.Component;
-import net.minecraft.world.MenuProvider;
-import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.AABB;
 import net.minecraftforge.network.PacketDistributor;
 
-public class UsernameLoggerBlockEntity extends DisguisableBlockEntity implements MenuProvider, ITickingBlockEntity, ILockable {
+public class UsernameLoggerBlockEntity extends DisguisableBlockEntity implements ITickingBlockEntity, ILockable {
 	private static final int TICKS_BETWEEN_ATTACKS = 80;
 	private IntOption searchRadius = new IntOption(this::getBlockPos, "searchRadius", 3, 1, 20, 1, true);
 	public String[] players = new String[100];
@@ -110,16 +105,6 @@ public class UsernameLoggerBlockEntity extends DisguisableBlockEntity implements
 
 	public void clearLoggedPlayersOnClient() {
 		SecurityCraft.channel.send(PacketDistributor.ALL.noArg(), new ClearLoggerClient(worldPosition));
-	}
-
-	@Override
-	public AbstractContainerMenu createMenu(int windowId, Inventory inv, Player player) {
-		return new GenericBEMenu(SCContent.USERNAME_LOGGER_MENU.get(), windowId, level, worldPosition);
-	}
-
-	@Override
-	public Component getDisplayName() {
-		return super.getDisplayName();
 	}
 
 	@Override