+++ src/main/java/net/geforcemods/securitycraft/blocks/mines/IMSBlock.java
@@ -4,10 +4,12 @@ import java.util.Random;
 
 import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.blockentities.IMSBlockEntity;
 import net.geforcemods.securitycraft.blocks.OwnableBlock;
+import net.geforcemods.securitycraft.network.client.OpenIMSScreen;
 import net.geforcemods.securitycraft.util.LevelUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -15,7 +17,6 @@ import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.MenuProvider;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
@@ -32,7 +33,7 @@ import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
-import net.minecraftforge.network.NetworkHooks;
+import net.minecraftforge.network.PacketDistributor;
 
 public class IMSBlock extends OwnableBlock {
 	public static final IntegerProperty MINES = IntegerProperty.create("mines", 0, 4);
@@ -105,10 +106,8 @@ public class IMSBlock extends OwnableBlock {
 					level.setBlockAndUpdate(pos, state.setValue(MINES, mines + 1));
 					((IMSBlockEntity) be).setBombsRemaining(mines + 1);
 				}
-				else if (player instanceof ServerPlayer) {
-					if (be instanceof MenuProvider menuProvider)
-						NetworkHooks.openGui((ServerPlayer) player, menuProvider, pos);
-				}
+				else
+					SecurityCraft.channel.send(PacketDistributor.PLAYER.with(() -> (ServerPlayer) player), new OpenIMSScreen(pos));
 			}
 		}