+++ Changelog.md
@@ -3,6 +3,7 @@
 - New: Block Change Detector. Detects and logs players breaking and/or placing blocks in its vicinity
 - Fix: The recipe for reinforced glass panes is incorrect
 - Fix: Breaking the block another block with modules is attached to does not drop the modules
+- Removed: Some unnecessary menu types. This may result in a "missing registry entries" message showing up, which can be accepted
 - Misc.: The minimum Forge version is now 40.0.35
 
 --------------------------Changelog for v1.9.1 of SecurityCraft--------------------------